IDE-1249: Initital draft of Spring Data integration.
Added support for Spring Data repositories, esp. the JPA and MongoDB modules. Repository interfaces are now marked as Spring beans in the navigator including a reference to the namespace element. The label and content in the Spring Beans tree have been shaped up accordingly. Beyond that there's a grouping node containing the Spring Data repository beans. The property elements in the XML editor should now be able to reference repository beans correctly as the bean instances in the model expose the repository interface as primary type.

The derived query methods in repository interfaces are validated semantically so that typos in property references and keywords are checked. Beyond that we provide property name code completion as well as proposals for keywords depending on on the just referenced property type.

The commit contains the plugin project as well as one containing some tests for the basic functionality:

TODOs:
- The bundles folder contains a Spring Data Commons JAR for the build. The repository metadata has to be built to include this dependency and allow it to be resolved during the Maven build.
- We need a feature project to wrap up the plugin and its dependencies