Make logger-name CA provider more reliable
This is achieved by doing smarter caching of results and making use
of cached results for a 'prefix' of the current query to answer current
query much faster.

We also made the time 'window' allowed for answering a query in UI thread
larger (so results come back more reliably). This is more acceptable now
than it was before, because this lengthy delay is typicially not happening
on every keystroke when user types a sequence of characters (thanks
to speedup of successive queries).