Fix PT-154181609
Convert .properties to .yml uses active editor as target even if 
invoked from a menu which makes a selection in the package explorer (or
other tree-viewer). If the active editor is not consistent with the tree viewer
selection this results in applying the conversion to another file than the
one selected by the user.

The fix is to make a more complex logic of determining the refactoring
which takes the 'active menu selection' into account as well.