Add Starters - Next button occasionally gets enabled on wrong URL
As a user types a wrong URL, the Next button occasionally gets enabled.
It's because the composite validator in the wizard sets OK as default
status while computing child validator values, sometimes the resulting
value is OK even if children are not OK. To solve this, just use the
wizard model validator directly to determine if Next should be enabled.