Add hk2 dependency back to boot.wizard plugin.
It is needed for jersey client 2.30, which in turn is a requirment
for mandas docker client. While boot.wizard does not need the docker
client. It makes sense to avoid installing multiple versions of
jersey client at the same time. Therefore we upgraded it now
to require 2.30 and so hk2 dependency is also required.